Rename file containing page definitions
